Rename LocalHistorStorage into VolatileHistoryStorage and fix a warning
