+++ test/CMakeLists.txt
@@ -42,7 +42,7 @@ add_executable(
 	test_suite
     driver.cpp
 	test_history.cpp
-	test_localhistorystorage.cpp
+	test_volatilehistorystorage.cpp
 	test_filehistorystorage.cpp
 	test_split.cpp
 	test_commonprefix.cpp
